requirements.txt
faiss-cpu
sentence-transformers
langchain-community
langchain
pypdf
streamlit
google-generativeai